Fix for renamed directory
