+++ tests/unit/schema/v1/test_template_builder.py
@@ -7,10 +7,10 @@ import pytest
 from mdio.core.v1.builder import MDIODatasetBuilder
 from mdio.core.v1.builder import _BuilderState
 from mdio.core.v1.builder import write_mdio_metadata
-from mdio.schema.compressors import Blosc
-from mdio.schema.dtype import ScalarType
-from mdio.schema.dtype import StructuredType
-from mdio.schema.v1.dataset import Dataset
+from mdio.schemas.compressors import Blosc
+from mdio.schemas.dtype import ScalarType
+from mdio.schemas.dtype import StructuredType
+from mdio.schemas.v1.dataset import Dataset
 
 
 def test_builder_initialization():
