+++ tests/unit/test_template_factory.py
@@ -16,9 +16,9 @@ from mdio.core.v1._serializer import make_variable
 from mdio.core.v1.builder import write_mdio_metadata
 from mdio.core.v1.factory import SCHEMA_TEMPLATE_MAP
 from mdio.core.v1.factory import MDIOSchemaType
-from mdio.schema.compressors import Blosc
-from mdio.schema.dtype import ScalarType
-from mdio.schema.dtype import StructuredType
+from mdio.schemas.compressors import Blosc
+from mdio.schemas.dtype import ScalarType
+from mdio.schemas.dtype import StructuredType
 
 
 def test_make_toy_dataset():
